requirements.txt
# DefineX 核心依赖
# 此文件仅包含运行 DefineX 所需的最小依赖集
# 完整依赖请查看 setup.py 中的 extras_require

# 基础框架
PyYAML>=6.0.1
rich>=13.7.1
click>=8.1.8

# MCP 协议支持
mcp>=1.0.0
fastmcp>=2.14.1

# Web 服务
uvicorn>=0.30.1
starlette>=0.47.2
sse-starlette>=2.2.1

# 安全
cryptography>=42.0.0

# 文件监控
watchdog>=4.0.0

# 网络请求
requests>=2.32.4

# 类型支持
typing-extensions>=4.0.0

# 终端优化
prompt-toolkit>=3.0.0

# OpenAI 支持
openai>=1.61.1

# 调试支持
httpx>=0.27.0
websockets>=12.0
